bsp: fix unused variable/function, sprintf overlap, IS_ERR_VALUE, and array-bounds warnings
cmsis-dap:
- DAP_config.h: mark 4 static functions __maybe_unused
- DAP.h: mark 2 static functions __maybe_unused, add #ifndef guards
  for DAP_SWD/DAP_JTAG macros (2 -Wundef)
- SWD_main.c: remove 7 unused variables
- swd_host.c: remove unused variable, fix declaration-after-statement

analog.c: use temp buffer + strlcat for snprintf overlap (-Wrestrict)
phy_device.c: use len+=sprintf pattern, add GENPHY_DRV_RTL8211F enum
  entry for excess-elements warning
xradio_btlpm.c, sunxi-wlan.c: cast to (long) before IS_ERR_VALUE
cpuinfo.c: remove unused serial variable
sunxi_timer_alarm.c: remove 2 unused flags variables
gc02m1_mipi.c: pragma-suppress false-positive -Warray-bounds

+++ drivers/bluetooth/xradio_btlpm.c
@@ -734,7 +734,7 @@ static int bluesleep_probe(struct platform_device *pdev)
 	//2.get bt_host_wake gpio irq
 #ifdef AW1732_BT
 	bsi->host_wake_irq = gpio_to_irq(bsi->host_wake);
-	if (IS_ERR_VALUE(bsi->host_wake_irq)) {
+	if (IS_ERR_VALUE((long)bsi->host_wake_irq)) {
 		BT_ERR("map gpio [%d] to virq failed, errno = %d\n", bsi->host_wake, bsi->host_wake_irq);
 		ret = -ENODEV;
 		goto free_bt_ext_wake;
